Add browser elements
Added browser elements to enable the game to be played by clicking in
the browser.  Created buttons for new game and reset game.  The control
buttons are hidden when not required.  The game board is locked until a
new game is started.

Need to add forms for entering player names and selecting a token.
Console messages regarding game state, winner and invalid moves also
need to be displayed in the browser

[Ticket: X]