Add some game messages
Area under gameboard now says if a player attempts an invalid move, if
the game is a draw and the winner's name.  Added player detail forms
to allow players to enter names and chosen tokens.  Player forms change
color to indicate the current player's turn.

Reset button doesn't change highlighted player to player 1 if pressed
when it is player 2's turn.